+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - FIXED: whois.ua should support status `ok`.
 
+- FIXED: whois.registry.net.za crashes when the registrant doesn't have a name.
+
 
 ## Release 3.0.0